Zero-Shot-Anomaly-Detection/requirements.txt
addict==2.4.0
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.4.0
async-lru==2.0.4
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.3
comm==0.1.4
cycler==0.10.0
debugpy==1.6.7.post1
decorator==4.4.2
defusedxml==0.7.1
exceptiongroup==1.1.3
executing==1.2.0
fairscale==0.4.4
fastjsonschema==2.18.0
filelock==3.12.0
fqdn==1.5.1
fsspec==2023.4.0
huggingface-hub==0.14.1
idna==2.10
importlib-metadata==6.8.0
importlib-resources==6.0.1
ipdb==0.13.4
ipykernel==6.25.2
ipython==8.12.2
ipython-genutils==0.2.0
ipywidgets==8.1.0
isoduration==20.11.0
jedi==0.18.0
Jinja2==3.1.2
joblib==1.0.0
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.24.0
kiwisolver==1.4.4
lit==16.0.3
MarkupSafe==2.1.2
matplotlib==3.1.3
matplotlib-inline==0.1.6
mistune==3.0.1
mpmath==1.3.0
natsort==8.4.0
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.7
notebook==7.0.3
notebook_shim==0.2.3
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-python==4.7.0.72
overrides==7.4.0
packaging==23.1
pandocfilters==1.5.0
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
platformdirs==3.10.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycocoevalcap==1.2
pycocotools==2.0.6
pycparser==2.21
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.0
referencing==0.30.2
regex==2023.5.5
requests==2.30.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
roboflow==1.1.4
rpds-py==0.10.2
ruamel.yaml==0.17.24
ruamel.yaml.clib==0.2.7
sacremoses==0.0.53
scikit-learn==0.24.1
Send2Trash==1.8.2
six==1.16.0
smplx==0.1.13
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.2
supervision==0.6.0
sympy==1.11.1
terminado==0.17.1
threadpoolctl==2.1.0
timm==0.4.12
tinycss2==1.2.1
tokenizers==0.10.3
tomli==2.0.1
torch==1.8.1+cu111
torchaudio==0.8.1
torchgeometry==0.1.2
torchvision==0.9.1+cu111
tornado==6.3.3
tqdm==4.65.0
traitlets==5.9.0
transformers==4.15.0
triton==2.0.0
typing_extensions==4.5.0
uri-template==1.3.0
urllib3==2.0.2
wcwidth==0.2.5
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.2
wget==3.2
widgetsnbextension==4.0.8
yapf==0.40.1
zipp==3.16.2
